refactor: reorder imports, add index files, replace long imports with paths/aliases
