+++ src/LCU/types/index.ts
@@ -2,3 +2,4 @@ export * from "./champions";
 export * from "./champselect";
 export * from "./lcucommon";
 export * from "./spells";
+export * from "./custom-mode";
