+++ src/components/lobbys/custom-lobby/summoner-in-lobby.tsx
@@ -1,7 +1,6 @@
-import { View, Button } from "@nodegui/react-nodegui";
 import React, { useContext } from "react";
-import { LCUContext } from "../../../LCU/lcucontext";
-import { LobbyMember } from "../../../LCU/types";
+import { View, Button } from "@nodegui/react-nodegui";
+import { LCUContext, LobbyMember } from "@lcu";
 import { LeaderPlayersActions } from "../leader-players-actions";
 
 interface SummonerInLobbyProps {
