+++ src/components/lobbys/default-queue/lobby-members.tsx
@@ -1,7 +1,6 @@
-import { Text, View } from "@nodegui/react-nodegui";
 import React, { useContext } from "react";
-import { LCUContext } from "../../../LCU/lcucontext";
-import { PositionsPreferences } from "../../../LCU/types";
+import { Text, View } from "@nodegui/react-nodegui";
+import { PositionsPreferences, LCUContext } from "@lcu";
 import { LeaderMemberManageActions } from "./leader-member-manage-actions";
 
 export function LobbyMembers() {
