+++ src/components/lobbys/leader-players-actions.tsx
@@ -1,6 +1,6 @@
-import { Button } from "@nodegui/react-nodegui";
 import React from "react";
-import { lcuClientHandlerObj } from "../../LCU/LCUClientHandler";
+import { Button } from "@nodegui/react-nodegui";
+import { lcuClientHandlerObj } from "@lcu";
 
 interface LeaderPlayersActionsProps {
   summonerId: number;
