+++ src/components/lobbys/lobby-actions.tsx
@@ -1,7 +1,6 @@
-import { Button, Text, View } from "@nodegui/react-nodegui";
 import React, { useContext, useState } from "react";
-import { LCUContext } from "../../LCU/lcucontext";
-import { queues } from "../../LCU/queues";
+import { Button, Text, View } from "@nodegui/react-nodegui";
+import { LCUContext, queues } from "@lcu";
 import { LeaveLobbyBtn } from "./leave-lobby-btn";
 import { DefaultQueue } from "./default-queue";
 import { textStyle } from "./styles";
