+++ src/components/phaseview/menu-client.tsx
@@ -1,8 +1,8 @@
 import { Text } from "@nodegui/react-nodegui";
 import React, { useContext } from "react";
-import { defaultTextStyle } from "./styles";
-import { LCUContext } from "../LCU/lcucontext";
-import { Lobbys } from "./lobbys";
+import { defaultTextStyle } from "../styles";
+import { LCUContext } from "@lcu";
+import { Lobbys } from "@components";
 
 export function MenuClient() {
   const { currentSummoner } = useContext(LCUContext);
